Fix: Use ansible_lsb.codename to get release name
- Fixes import key error which was initially hardcoded:

E: repo 'http://apt.puppetlabs.com xenial InRelease' is not signed
W:Updating from such a repository can't be done securely, and is
therefore disabled by default.

- Update molecule config to fix:

line 8, column 7, but may be elsewhere in the file depending on
the exact syntax problem.

  The offending line appears to be:

    roles:
      - "{{ dynamic_role_name }}"
        ^ here

This tells Molecule/Ansible to use the current role (i.e., the one
being tested). This is the canonical way to avoid all the
lookup('env', ...) gymnastics — and it works cleanly even
during --syntax-check.

Change-Id: I42b335530a5dbb6ee9339b5e56992ece23b9138f
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>